better unittests
