+++ test-js/test-var-glyph.js
@@ -2,6 +2,7 @@ import { expect } from "chai";
 
 import { enumerate } from "../src/fontra/client/core/utils.js";
 import { StaticGlyph, VariableGlyph } from "../src/fontra/client/core/var-glyph.js";
+import { dummyStaticGlyph } from "./test-globals.js";
 
 function makeTestGlyphObject() {
   return {
@@ -101,39 +102,38 @@ describe("var-glyph Tests", () => {
     expect(glyph).to.deep.equal(denseObject);
   });
 
-  it("modify StaticGlyph leftMargin", () => {
-    const sparseObject = {
-      xAdvance: 170,
-      leftMargin: 60,
-      rightMargin: 60,
-      path: {
-        contourInfo: [{ endPoint: 3, isClosed: true }],
-        coordinates: [60, 0, 110, 0, 110, 120, 60, 120],
-        pointTypes: [0, 0, 0, 0],
-      },
-    };
-    const glyph = StaticGlyph.fromObject(sparseObject);
-    glyph.leftMargin += 10;
-    expect(glyph.xAdvance).to.equal(180);
-    expect(glyph.leftMargin).to.equal(70);
-    expect(glyph.coordinates).to.equal([70, 0, 120, 0, 120, 120, 70, 120]);
+  it("modify StaticGlyph leftMargin check xAdvance", () => {
+    dummyStaticGlyph.leftMargin += 10;
+    expect(dummyStaticGlyph.xAdvance).to.equal(180);
   });
 
-  it("modify StaticGlyph rightMargin", () => {
-    const sparseObject = {
-      xAdvance: 170,
-      leftMargin: 60,
-      rightMargin: 60,
-      path: {
-        contourInfo: [{ endPoint: 3, isClosed: true }],
-        coordinates: [60, 0, 110, 0, 110, 120, 60, 120],
-        pointTypes: [0, 0, 0, 0],
-      },
-    };
-    const glyph = StaticGlyph.fromObject(sparseObject);
-    glyph.rightMargin += 10;
-    expect(glyph.xAdvance).to.equal(180);
-    expect(glyph.rightMargin).to.equal(70);
+  it("modify StaticGlyph leftMargin check leftMargin", () => {
+    dummyStaticGlyph.leftMargin += 10;
+    expect(dummyStaticGlyph.leftMargin).to.equal(70);
+  });
+
+  it("modify StaticGlyph leftMargin check coordinates", () => {
+    dummyStaticGlyph.leftMargin += 10;
+    expect(dummyStaticGlyph.path.coordinates).to.equal([
+      70, 0, 120, 0, 120, 120, 70, 120,
+    ]);
+  });
+
+  it("modify StaticGlyph rightMargin check xAdvance", () => {
+    dummyStaticGlyph.rightMargin += 10;
+    expect(dummyStaticGlyph.xAdvance).to.equal(180);
+  });
+
+  it("modify StaticGlyph rightMargin check leftMargin", () => {
+    dummyStaticGlyph.rightMargin += 10;
+    expect(dummyStaticGlyph.rightMargin).to.equal(70);
+  });
+
+  it("modify StaticGlyph rightMargin check coordinates", () => {
+    dummyStaticGlyph.rightMargin += 10;
+    expect(dummyStaticGlyph.path.coordinates).to.equal([
+      60, 0, 110, 0, 110, 120, 60, 120,
+    ]);
   });
 
   const modifierFuncs = [
